Record error if no prior daily is found but don't rethrow
When running the daily jobs, primary data is being generated but there
is no prior JSON data to compare against, so an error is thrown. What we
should be doing is recording the error but letting the build proceed so
that the primary results are saved and be compared against in the next
daily run.