Fix the incorrect use of redirect_std* functions (#62)
The code as implemented in #61 assumes that `redirect_std*` returns a
tuple. That is only true when the function is given no arguments. When
given a single `IOStream` as an argument, the function does a pointer
store then returns the input stream. That causes an error when using the
destructuring assignment syntax (no method matching
`iterate(::IOStream)`).

This change mostly reverts #61, except it keeps the redirects back to
the original STDOUT/STDERR streams.